fix: Remove obsolete python version 3.7 from workflow matrix
Workflow CI tests were launched on a matrix of python versions: 3.7, 3.8 and 3.9. Recently, workflow builds on python 3.7 have started to fail but we probably do not want to invest any efforts on trying to fix a build for an obsolete python version. Henceforth, this commit removes the problematic version from the CI workflow matrix.